// frontend/hooks/models/songDTOVerificationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type SongDTOVerificationStatus = typeof SongDTOVerificationStatus[keyof typeof SongDTOVerificationStatus];


export const SongDTOVerificationStatus = {
  UNVERIFIED: 'UNVERIFIED',
  VERIFIED: 'VERIFIED',
  NEEDS_REVIEW: 'NEEDS_REVIEW',
  MANUAL_ENTRY: 'MANUAL_ENTRY',
} as const;